Add "constrain to pirmary axis" functionality
I still haven't figured out exactly what I need to do to get the control that I want. It seems that the horizontal rotation matrix is imparting some undesirable rotation into the vertical rotation matrix, and into the camera's upVector. I don't know how I could make it work such that I could not have it affect camera/vertical rotation, but also keep track of it so I can properly change the upvector when it needs to happen (and not when it's unwanted).

For now, I've added some peices that get close to functioning how I want:
 - Fixed getConstrained, the original code was taken from the Fusion 360 sample program and seems to have a bug, it'd always return the same vector (also, it's kinda wildly inefficient). Now I just find the biggest axis, and assume that's the closest primary axis. This operates how I'd expect, and is nice and quick/easy computation.
 - Constrain the verticalRotationMatrix to a primary axis. This makes it much easier to rotate only vertically, and not end up doing the odd vertical+horizontal rotation that's unwanted.
 - Added a button command to snap the upVector to a primary axis. This makes it so you can easily "untilt" the view, which is the largest source of annoyance for me.

 I don't think this is a proper final solution, and it'd probably be better to build in tilt controls to do this more naturally, but the Xbox gamepad only has 2.5 joysticks, so there's not really anywhere to put a tilt axis (and doing it in some increment with a button press wouldn't really feel good imo).