Update to utilize pygame library
`pyjoystick` "works" for the xbox controller, but when I tried to use my new custom built controller, it seemed to not be getting the data from the joystick (even though it shows up in windows).

Updating the add-in to use a dynamically installed `pygame`. If the (hacky) installation of `pygame` fails, we should be able to fall back to `pyjoystick`.